+++ gateway/gateway_test.go
@@ -1,22 +1,23 @@
-package main
+package gateway
 
 import (
 	"testing"
 	"encoding/base64"
 	"log"
 	"encoding/json"
+	"github.com/Ma-Shell/Holmes-Gateway/utils"
 	)
 
 func TestRSA(t *testing.T) {
 	print("RSA-Test\n")
 	rsakey,_ := loadPrivateKey("keys/blub.priv")
 	p1 := "abcdef0123456789"
-	c, err := RsaEncrypt([]byte(p1), &rsakey.PublicKey)
+	c, err := tasking.RsaEncrypt([]byte(p1), &rsakey.PublicKey)
 	if err != nil {
 		t.Error(err)
 	}
 	print(base64.StdEncoding.EncodeToString(c)+"\n")
-	p2, err := RsaDecrypt(c, rsakey)
+	p2, err := tasking.RsaDecrypt(c, rsakey)
 	if err != nil {
 		t.Error(err)
 	} else if (p1 != string(p2)) {
@@ -25,7 +26,7 @@ func TestRSA(t *testing.T) {
 }
 
 func TestValidateTask(t *testing.T) {
-	ta := Task{
+	ta := tasking.Task{
 		PrimaryURI : "http://127.0.0.1:8016/samples/3a12f43eeb0c45d241a8f447d4661d9746d6ea35990953334f5ec675f60e36c5",
 		SecondaryURI : "",
 		Filename : "myfile",
@@ -33,7 +34,7 @@ func TestValidateTask(t *testing.T) {
 		Tags : []string{"test1"},
 		Attempts : 0 }
 
-	task, err := json.Marshal([]Task{ta})
+	task, err := json.Marshal([]tasking.Task{ta})
 	if err != nil {
 		t.Error(err)
 	}
@@ -52,7 +53,7 @@ func TestAESDecrypt(t *testing.T) {
 		t.Error(err)
 	}
 
-	v, err := AesDecrypt(ciphertext, []byte("abcdef0123456789"), []byte("0000111122223333"))
+	v, err := tasking.AesDecrypt(ciphertext, []byte("abcdef0123456789"), []byte("0000111122223333"))
 	if err != nil {
 		t.Error(err)
 	} else if (string(v) != ("test encryption!")) {
@@ -62,7 +63,7 @@ func TestAESDecrypt(t *testing.T) {
 
 func TestAES(t *testing.T) {
 	print("AES-Test\n")
-	ta := Task{
+	ta := tasking.Task{
 		PrimaryURI : "http://127.0.0.1:8016/samples/3a12f43eeb0c45d241a8f447d4661d9746d6ea35990953334f5ec675f60e36c5",
 		SecondaryURI : "",
 		Filename : "myfile",
@@ -70,7 +71,7 @@ func TestAES(t *testing.T) {
 		Tags : []string{"test1"},
 		Attempts : 5 }
 
-	task, err := json.Marshal([]Task{ta})
+	task, err := json.Marshal([]tasking.Task{ta})
 	if err != nil {
 		t.Error(err)
 	}
@@ -80,14 +81,14 @@ func TestAES(t *testing.T) {
 	key := []byte("abcdef0123456789")
 	iv := []byte("0000111122223333")
 
-	ciphertext, err := AesEncrypt(plaintext, key, iv)
+	ciphertext, err := tasking.AesEncrypt(plaintext, key, iv)
 
 	if err != nil {
 		t.Error(err)
 	}
 	print(base64.StdEncoding.EncodeToString(ciphertext)+"\n")
 
-	plaintext2, err := AesDecrypt(ciphertext, key, iv)
+	plaintext2, err := tasking.AesDecrypt(ciphertext, key, iv)
 	if err != nil {
 		t.Error(err)
 	} else if (string(plaintext2) != string(plaintext)) {
